+++ examples/package.json
@@ -56,6 +56,7 @@
     "@pinecone-database/pinecone": "^4.0.0",
     "@llamaindex/assemblyai": "^0.1.0",
     "@llamaindex/discord": "^0.1.0",
+    "@llamaindex/notion": "^0.1.0",
     "@vercel/postgres": "^0.10.0",
     "ai": "^4.0.0",
     "ajv": "^8.17.1",
